Cast the hint count for Debug builds
With -Wconversion in the baseline, GCC reports the size_t hint count
passed to dice() in a Debug build, where it is not folded away as it is
in the optimized builds the budget used. The CMake Debug jobs failed on
it under -Werror; Debug baseline builds are now clean on GCC 13, GCC 16,
Clang 18, and Clang 22.

+++ src/act.other.c
@@ -10464,7 +10464,7 @@ static void show_hint_index(struct char_data *ch, int roll)
 
 void show_hint_one(struct char_data *ch)
 {
-  show_hint_index(ch, dice(1, NUM_HINTS) - 1);
+  show_hint_index(ch, dice(1, (int)NUM_HINTS) - 1);
 }
